S1C7 - Hand improvements and Deck implementation
When drawing card they now enter a queue which means that the correct card is drawn but only when it is possible for the player to draw another card allowing as many card draws to occur at once and all still be handled correctly. The deck has been added with the functionality to shuffle it using the The Fisher–Yates shuffle algorithm. This ability is maintained within Unity's inspector.