S1C11 - Adding cards to the stack
You can now drag a card from your hand onto the board. When you do this the card is then added to the stack. Once a card is told to resolve the type of the card is considered by a switch case. If the type is creature a prefab is spawned for use on the board.