S2C1 - Board Formatting and Phases
When the stack is not enabled the user can now choose the position that the card is placed. This is done by getting the mouse position and comparing it to the other cards already on the board. Phases have also been introduced, the user can be in different phases and the phases can currently be progressed through using a button.