S2C2 -  Targeting line
A line is now drawn from the player icon to the position of the mouse to represent where the target will be. This only happens with spells and not creatures. there is an issue with spells still having attack and health values. Different card frames can be used using a struct in the inspector.